requirements.txt
streamlit==1.36.0
google-generativeai==0.8.2